Refactoring the way cards are displayed
