+++ frontend/src/App.scss
@@ -18,6 +18,7 @@ nav {
 }
 
 #search {
+  position: relative;
   max-width: 30rem;
   margin: 3rem auto;
   display: flex;
@@ -25,30 +26,51 @@ nav {
   .btn {
     width: 30%;
   }
+
+  small {
+    position: absolute;
+    top: 3rem;
+  }
 }
 
 #cards-wrapper {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-around;
-
-  .card {
-    padding: 1rem;
-    width: 20%;
+  img {
+    max-width: 100%;
+    display: block;
+  }
+  
+  figure {
+    margin: 0;
+    display: grid;
+    grid-template-rows: 1fr auto;
+    margin-bottom: 10px;
+    break-inside: avoid;
+  }
+  
+  figure > img {
+    grid-row: 1 / -1;
+    grid-column: 1;
   }
 
-  .card-body {
-    position: relative;
+  figure span {
+    color: black;
+    text-decoration: none;
+    cursor: pointer;
+  }
 
-    .card-title {
-      margin: 0 0 2rem;
-    }
+  figcaption {
+    grid-row: 2;
+    grid-column: 1;
+    background-color: rgba(255,255,255,.5);
+    padding: .2em .5em;
+    justify-self: start;
+  }
 
-    button {
-      position: absolute;
-      bottom: 5px;
-    }
-  } 
+  .container {
+    column-count: 4;
+    column-gap: 10px;
+    width: 100%;
+  }
 }
 
 .empty {
